Separated LP class from training script
